+++ packages/protocol/test/rrp/TemplateUtils.sol.js
@@ -2,7 +2,7 @@
 
 const hre = require('hardhat');
 const { expect } = require('chai');
-const utils = require('./utils');
+const utils = require('../utils');
 
 let roles;
 let airnodeRrp;
