+++ packages/protocol/test/rrp/authorizers/Api3RequesterRrpAuthorizer.sol.js
@@ -2,7 +2,7 @@
 
 const hre = require('hardhat');
 const { expect } = require('chai');
-const utils = require('../utils');
+const utils = require('../../utils');
 
 const AdminRank = Object.freeze({
   Unauthorized: 0,
